Improving structure of Queue programs
Making a common data-structure of Queue. After this commit, all the
solutions of Queue should depend on this data structure and add
functionality according to their need.
